update unity client sample with massive refactoring

+++ src/shared/UdpToolkit.Core/TimerKey.cs
@@ -3,14 +3,14 @@ namespace UdpToolkit.Core
     public readonly struct TimerKey
     {
         public TimerKey(
-            ushort roomId,
+            int roomId,
             short timerId)
         {
             RoomId = roomId;
             TimerId = timerId;
         }
 
-        public ushort RoomId { get; }
+        public int RoomId { get; }
 
         public short TimerId { get; }
     }
